requirements.txt
asgiref==3.10.0
Django==5.2.7
python-dotenv==1.1.1
sqlparse==0.5.3

# Django REST framework (for API and token auth)
djangorestframework==3.15.0
